+++ src/test/scala/io/viash/platforms/nextflow/NextflowScriptTest.scala
@@ -80,6 +80,22 @@ class NextflowScriptTest extends AnyFunSuite with BeforeAndAfterAll {
     assert(exitCode == 0, s"\nexit code was $exitCode\nStd output:\n$stdOut\nStd error:\n$stdErr")
   }
 
+  test("Run config pipeline with subworkflow dependency and symlinks", NextflowTest) {
+    val newWorkflowPath = Paths.get(tempFolStr, "nestedWorkflowsAsSymlink")
+    Files.createDirectories(newWorkflowPath)
+    val symlinkFolder = Paths.get(newWorkflowPath.toString(), "workflow")
+    Files.createSymbolicLink(symlinkFolder, Paths.get(tempFolStr, "target/nextflow/test_wfs/"))
+    val (exitCode, stdOut, stdErr) = NextflowTestHelper.run(
+      mainScript = "nestedWorkflowsAsSymlink/workflow/nested/main.nf",
+      args = List(
+        "--publish_dir", "output"
+      ),
+      cwd = tempFolFile
+    )
+
+    assert(exitCode == 0, s"\nexit code was $exitCode\nStd output:\n$stdOut\nStd error:\n$stdErr")
+  }
+
   // TODO: use TestHelper.testMainWithStdErr instead of NextflowTestHelper.run; i.e. viash test
   test("Test workflow", DockerTest, NextflowTest) {
     val (exitCode, stdOut, stdErr) = NextflowTestHelper.run(